interface admin

+++ view/user/Admin/code.php
@@ -74,11 +74,36 @@ if (isset($_POST['update_Admin_btn'])){
     }
 }
 
+// Update Seller profile
+if (isset($_POST['update_Seller_btn'])){
+    $id=$_POST['edit_id'];
+    $name=$_POST['edit_username'];
+    $username = $_POST['edit_pseudo'];
+    $email = $_POST['edit_email'];
+    $password = $_POST['edit_password'];
+
+    $query="UPDATE seller SET name='$name',username='username',email='$email', password='$password' where id='$id'";
+    $query_run=mysqli_query($connection,$query);
+
+    if ($query_run){
+        session_start();
+        $_SESSION['success']= "Your data is update";
+        header('location:registerSeller.php');
+        exit();
+    }else{
+        session_start();
+        $_SESSION['status']= "Your data has not update";
+        header('location:registerSeller.php');
+        exit();
+    }
+}
+
 // Delete Admin profile
-if (isset($_POST['delete_btn'])){
-    echo ("test");
+if (isset($_POST['delete_admin_btn'])){
     $id=$_POST['delete_id'];
-    $query="DELETE FROM admin WHERE id='$id'";
+
+    $query="DELETE FROM `admin` WHERE `admin`.`id` ='$id'";
+
     $query_run=mysqli_query($connection,$query);
     if ($query_run){
         $_SESSION['success']= "Your data is DELETE";
@@ -86,6 +111,38 @@ if (isset($_POST['delete_btn'])){
     }else{
         $_SESSION['status']= "Your data has not DELETE";
         header('location:registerAdmin.php');
+    }
+}
+
+// Delete Seller profile
+if (isset($_POST['delete_seller_btn'])){
+    $id=$_POST['delete_id'];
+
+    $query="DELETE FROM `seller` WHERE `seller`.`id` ='$id'";
+
+    $query_run=mysqli_query($connection,$query);
+    if ($query_run){
+        $_SESSION['success']= "Your data is DELETE";
+        header('location:registerSeller.php');
+    }else{
+        $_SESSION['status']= "Your data has not DELETE";
+        header('location:registerSeller.php');
+    }
+}
+
+// Delete Seller profile
+if (isset($_POST['delete_user_btn'])){
+    $id=$_POST['delete_id'];
+
+    $query="DELETE FROM `customer` WHERE `customer`.`id` ='$id'";
+
+    $query_run=mysqli_query($connection,$query);
+    if ($query_run){
+        $_SESSION['success']= "Your data is DELETE";
+        header('location:registerUser.php');
+    }else{
+        $_SESSION['status']= "Your data has not DELETE";
+        header('location:registerUser.php');
     }
 }
 ?>
